backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
openai
python-dotenv
xlrd
openpyxl
